% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get data}
\usage{
get_data(location_id = NULL)
}
\arguments{
\item{location_id}{Unique ID of location.}
}
\value{
Data frame
}
\description{
Import data from web services (currently EA only) and convert into standard
format for `hera` tools.
}
\examples{
data <- get_data(location_id = 92751)
}
