% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafpacs-classify.R
\name{leafpacs_classify}
\alias{leafpacs_classify}
\title{Classify River LEAFPACS}
\usage{
leafpacs_classify(data)
}
\arguments{
\item{data}{Data frame of predicted and observed macrophyte indices}
}
\value{
Data frame containing classification results
}
\description{
Using pre-calculated sample metrics and predicted reference values this
function calculates the EQRs and classification for each sample.
}
\examples{
\dontrun{
class <- leafpacs_classify(data)
}
}
