% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafpacs-confidence.R
\name{leafpacs_confidence}
\alias{leafpacs_confidence}
\title{Calculate Confidence of Class}
\usage{
leafpacs_confidence(data)
}
\arguments{
\item{data}{Data frame of classification results}
}
\value{
Data frame containing confidence of class
}
\description{
Calculate Confidence of Class
}
\examples{
data <- data.frame(eqr = c(0.999))
confidence <- leafpacs_confidence(data)
}
