% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafpacs-convert-ea.R
\name{leafpacs_convert_ea}
\alias{leafpacs_convert_ea}
\title{Convert EA data to standard input}
\usage{
leafpacs_convert_ea(data)
}
\arguments{
\item{data}{Data frame of macrophyte taxa and cover values from EA bulk
download.}
}
\value{
Data frame
}
\description{
Convert EA data to standard input
}
\examples{
# Load data
library(dplyr)
indices <- utils::read.csv(system.file("extdat",
                                       "MACP_OPEN_DATA_METRICS.csv",
                                       package = "leafpacs"
))

predictors <- utils::read.csv(system.file("extdat",
                                          "MACP_OPEN_DATA_SITE.csv",
                                          package = "leafpacs"
))
# Tidy data
data <- dplyr::inner_join(indices, predictors, by = "SITE_ID")
data <- dplyr::select(data, -.data$REPLICATE_CODE)
data <- data \%>\% dplyr::filter(complete.cases(data))
data <- leafpacs_convert_ea(data)
}
