% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafpacs-indices.R
\name{leafpacs_indices}
\alias{leafpacs_indices}
\title{Calculate river LEAFPACS indices}
\usage{
leafpacs_indices(data)
}
\arguments{
\item{data}{Data frame of macrophyte taxa and cover values.}
}
\value{
Data frame containing indices
}
\description{
Calculate river LEAFPACS indices
}
\examples{
\dontrun{
indices <- leafpacs_indices(data)
}
}
