% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafpacs-predict.R
\name{leafpacs_predict}
\alias{leafpacs_predict}
\title{Predict LEAFPACS RMNI}
\usage{
leafpacs_predict(data, reference_algae = 0.05)
}
\arguments{
\item{data}{Data frame of river LEAFPACS predictors}

\item{reference_algae}{Number representing reference algae. Default is 0.05
set by the river LEAFPACS standard.}
}
\value{
Data frame of reference predictions
}
\description{
Function to predict LEAFPACS River Macrophyte Nutrient Index (RMNI), Number
of Functional Groups and add 0.05 for reference algae.
}
\examples{
\dontrun{
predictions <- leafpacs_predict(data)
}
}
