% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Validate}
\usage{
validate(data)
}
\arguments{
\item{data}{Data frame of macrophyte taxonomic data and/or predictors}
}
\value{
Data frame of validated input data
}
\description{
Validate input data for correctness.
}
\examples{
\dontrun{
data <- validate(1)
}
}
