% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafpacs.R
\name{leafpacs}
\alias{leafpacs}
\title{Run LEAFPACS indices, prediction and classification}
\usage{
leafpacs(data)
}
\arguments{
\item{data}{Data frame of taxonomic or indices results}
}
\value{
Data frame containing indices, predictions and classification
}
\description{
This is a wrapper function to run all the utility functions together to give
a final output.
}
\examples{
results <- leafpacs(taxa_data)
\dontrun{
data <- hera::get_data(location_id = 92751)
class <- leafpacs(data)
}
}
