% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafpacs.R
\name{leafpacs}
\alias{leafpacs}
\title{Run all LEAFPACS indices, predictions and classification}
\usage{
leafpacs(data)
}
\arguments{
\item{data}{Data frame of taxonomic or indices results}
}
\value{
Data frame containing indices, predictions and classification
}
\description{
Run all LEAFPACS indices, predictions and classification
}
\examples{
\dontrun{
results <- leafpacs(data)
}
}
