% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafpacs-confidence.R
\name{leafpacs_confidence}
\alias{leafpacs_confidence}
\title{Confidence of class}
\usage{
leafpacs_confidence(data)
}
\arguments{
\item{data}{Data frame of classification results}
}
\value{
Data frame containing confidence of class
}
\description{
Confidence of class
}
